pyproject.toml
[project]
name = "ocr-agent"
version = "0.1.0"
description = "Multi-agent OCR system with cascading fallback and beautiful terminal UI"
authors = [
    {name = "Ruben Fernandez-Fuertes", email = "fernandezfuertesruben@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["ocr", "multi-agent", "document-processing", "pdf", "nougat", "mistral", "gemini", "deepseek"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "rich>=13.0.0",
    "click>=8.1.0",
    "pillow>=10.0.0",
    "pymupdf>=1.24.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
]
# Optional: Install specific OCR engines
nougat = ["nougat-ocr>=0.1.17", "torch>=2.0.0"]
gemini = ["google-genai>=1.0.0"]
mistral = ["mistralai>=1.0.0"]
# Install all cloud engines
cloud = ["google-genai>=1.0.0", "mistralai>=1.0.0"]
# Install everything
all = ["nougat-ocr>=0.1.17", "torch>=2.0.0", "google-genai>=1.0.0", "mistralai>=1.0.0"]

[project.scripts]
ocr-agent = "ocr_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ocr_agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
